+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 ## [Unreleased] — 2026-07-18
 
 ### Changed
+- Merge pull request #19 from Asterion-Technology/feature/RAD-15-portal-actions
+
+
+## [Unreleased] — 2026-07-18
+
+### Changed
+- Merge pull request #16 from Asterion-Technology/feature/RAD-15-staff-portal
+
+
+## [Unreleased] — 2026-07-18
+
+### Changed
+- Merge pull request #18 from Asterion-Technology/feature/RAD-17-ca-tax
+
+
+## [Unreleased] — 2026-07-18
+
+### Changed
+- Merge pull request #15 from Asterion-Technology/feature/RAD-18-email-provider
+
+
+## [Unreleased] — 2026-07-18
+
+### Changed
 - Merge pull request #13 from Asterion-Technology/feature/RAD-15-operations